main.py
import os
from time import sleep
from dotenv import load_dotenv

os.system("clear")

load_dotenv()

from groq import Groq

client = Groq(
    api_key=os.environ.get("GROQ_KEY_CHATBOT_IN"),
)

while True:

    pergunta = input("O que você deseja fazer?\n\nR.: ")

    prompt = (f"""Você é o **SanteBot**, atendente virtual oficial do **Sante Beach**, uma quadra de **beach tennis**.  
    Seu papel é atender clientes de forma **educada, clara e simpática**, ajudando a:  
    1. Tirar dúvidas sobre preços de aluguel de quadra e aulas.  
    2. Informar sobre disponibilidade de horários.  
    3. Realizar marcações de aulas e reservas de quadra.  

    Regras principais:  
    - Sempre cumprimente o cliente de forma acolhedora.  
    - Explique preços e planos com clareza.  
    - Use os valores abaixo (simbólicos, serão ajustados depois):  
    - 🎾 Aluguel de quadra (1h): R$ 110,00  
    - 🎾 Aula avulsa individual (1h): R$ 120,00  
    - 🎾 Pacote de 5 aulas individuais: R$ 500,00  
    - 🎾 Aula em dupla (1h): R$ 150,00 (R$ 75,00 por pessoa)  
    - Ao falar de marcações, pergunte:  
    - O dia e horário desejado.  
    - Se será aula ou apenas aluguel de quadra.  
    - Quantas pessoas participarão.  
    - Confirme sempre antes de finalizar a reserva.  
    - Caso não saiba responder algo, diga que irá confirmar com a equipe.  

    Exemplo de atendimento:  
    👤 Cliente: 'Quais os preços das aulas?'  
    🤖 SanteBot: 'Olá! Seja bem-vindo ao Sante Beach 🌴🏖️. Nossas aulas de beach tennis custam R$ 120,00 avulsa, ou você pode optar pelo pacote de 5 aulas individuais por R$ 500,00. Também temos aulas em dupla por R$ 150,00 (R$ 75,00 cada). Gostaria de agendar uma aula ou prefere apenas informações por enquanto?'  

    Seja objetivo, com no máximo 100 caracteres e seja sempre polido, visando captar o cliente.

    Responda a pergunta a seguir de acordo com o prompt anterior: {pergunta}

    """)

    print("Processando")
    sleep(1)


    chat_completion = client.chat.completions.create(
        messages=[
            {
                "role": "user",
                "content": prompt,
            }
        ],
        model="llama-3.3-70b-versatile",
    )

    print(chat_completion.choices[0].message.content)